+++ scripts/create_registry_schema.py
@@ -149,6 +149,7 @@ def _get_table_metadata(schema, table):
                 UniqueConstraint(
                     *schema_data[table]["unique_constraints"][uq_att]["unique_list"],
                     name=uq_att,
+                    postgresql_nulls_not_distinct=True,
                 )
             )
 
@@ -284,9 +285,9 @@ def _DatasetKeyword(schema):
 # The following should be adjusted whenever there is a change to the structure
 # of the database tables.
 _DB_VERSION_MAJOR = 3
-_DB_VERSION_MINOR = 1
+_DB_VERSION_MINOR = 3
 _DB_VERSION_PATCH = 0
-_DB_VERSION_COMMENT = "Add dataset.access_api_configuration and dataset_alias.ref_alias_id"
+_DB_VERSION_COMMENT = "Remove `is_overwritten`, `replace_date` and `replace_uid` columns, the information is in `status`"
 
 # Parse command line arguments
 parser = argparse.ArgumentParser(
@@ -329,7 +330,6 @@ else:
             with db_connection.engine.connect() as conn:
                 result = conn.execute(text(stmt))
                 result = pd.DataFrame(result)
-                conn.commit()
         except Exception:
             raise RuntimeError("production schema does not exist or is ill-formed")
         if (
@@ -361,7 +361,7 @@ if schema:
                 select_priv = f"GRANT SELECT ON ALL TABLES IN SCHEMA {schema} to {acct}"
                 conn.execute(text(usage_priv))
                 conn.execute(text(select_priv))
-                conn.commit()
+            conn.commit()
     except Exception as e:
         print(f"Could not grant access to {acct} on schema {schema}")